Added the python file containing the full game
To play the game the LM studio llm needs to be running and listening and the python file needs to be run. There is only voice input. User needs to speak and introduce himself right after the ais did their introduction. then the Ais viode and the user needs to vote as well.